change courses repository directory
