+++ src/modules/courses/repositories/ICoursesRepository.ts
@@ -1,0 +1,14 @@
+import { Course } from '../../../entities/Course';
+import {
+  ICourseWithUsers,
+  IGetCourseWithUsersDTO,
+} from '../useCases/GetCourseWithUsers/GetCourseWithUsersDTO';
+import { IIndexCourseRequestDTO } from '../useCases/IndexCourse/IndexCourseDTO';
+import { IPaginatedArray } from '../../../providers/IArrayPaginatorProvider';
+
+export interface ICoursesRepository {
+  create(course: Course): Promise<void>;
+  index(data: IIndexCourseRequestDTO): Promise<IPaginatedArray>;
+
+  getUsers(data: IGetCourseWithUsersDTO): Promise<ICourseWithUsers[]>;
+}
